Update main.cpp.ino
Found out that to go forward, right servo must be 0 and left must be 180. Wrote the forward function that moves the buggy forward, and every millisecond checks on whether it's dark. If it's dark, it turns the lights on. If it's bright, it waits one second, then turns the lights off.